Add support for file overwriting on Local and Azure file services. Update packages.

+++ Services/Storage.LocalFile/LocalFileStorageService.cs
@@ -33,7 +33,7 @@ namespace EllipticBit.Services.Storage {
 			return Task.FromResult((Stream)File.OpenRead(cp));
 		}
 
-		public async Task Write(string path, Stream data) {
+		public async Task Write(string path, Stream data, bool overwrite = false) {
 			var cp = Path.Combine(rootPath, path);
 			var dir = Path.GetDirectoryName(cp);
 
@@ -41,7 +41,7 @@ namespace EllipticBit.Services.Storage {
 				Directory.CreateDirectory(dir);
 			}
 
-			using (var fs = File.OpenWrite(cp)) {
+			using (var fs = File.Open(cp, overwrite ? FileMode.Create : FileMode.CreateNew, FileAccess.Write)) {
 				await data.CopyToAsync(fs);
 			}
 		}
@@ -51,17 +51,25 @@ namespace EllipticBit.Services.Storage {
 			return Task.FromResult(File.Exists(cp));
 		}
 
-		public Task<bool> Move(string sourcePath, string targetPath) {
+		public Task<bool> Move(string sourcePath, string targetPath, bool overwrite = false) {
 			var src = Path.Combine(rootPath, sourcePath);
 			var tgt = Path.Combine(rootPath, targetPath);
+
+			if (!overwrite && File.Exists(tgt)) {
+				throw new IOException($"The specified target file already exists: {tgt}");
+			}
+			if (overwrite && File.Exists(tgt)) {
+				File.Delete(tgt);
+			}
+
 			File.Move(src, tgt);
 			return Task.FromResult(true);
 		}
 
-		public Task<bool> Copy(string sourcePath, string targetPath) {
+		public Task<bool> Copy(string sourcePath, string targetPath, bool overwrite = false) {
 			var src = Path.Combine(rootPath, sourcePath);
 			var tgt = Path.Combine(rootPath, targetPath);
-			File.Copy(src, tgt);
+			File.Copy(src, tgt, overwrite);
 			return Task.FromResult(true);
 		}
 	}
